src/requirements.txt
Django
djangorestframework
pymysql
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
cryptography
ultralytics
opencv-python
torchvision
numpy
pillow
requests
drf-yasg
whitenoise
gunicorn